removed example files, connected/downloaded axios, now have functioning index route
