--- app/models/example.js
@@ -1,24 +1,0 @@
-const mongoose = require('mongoose')
-
-const exampleSchema = new mongoose.Schema(
-	{
-		title: {
-			type: String,
-			required: true,
-		},
-		text: {
-			type: String,
-			required: true,
-		},
-		owner: {
-			type: mongoose.Schema.Types.ObjectId,
-			ref: 'User',
-			required: true,
-		},
-	},
-	{
-		timestamps: true,
-	}
-)
-
-module.exports = mongoose.model('Example', exampleSchema)
