+++ app/routes/activity_routes.js
@@ -1,0 +1,73 @@
+const express = require('express')
+const passport = require('passport')
+const axios = require('axios').default
+
+// pull in Mongoose model for activity
+const Activity = require('../models/activity')
+
+// this is a collection of methods that help us detect situations when we need
+// to throw a custom error
+const customErrors = require('../../lib/custom_errors')
+
+// we'll use this function to send 404 when non-existant document is requested
+const handle404 = customErrors.handle404
+// we'll use this function to send 401 when a user tries to modify a resource
+// that's owned by someone else
+const requireOwnership = customErrors.requireOwnership
+
+// this is middleware that will remove blank fields from `req.body`, e.g.
+// { example: { title: '', text: 'foo' } } -> { example: { text: 'foo' } }
+const removeBlanks = require('../../lib/remove_blank_fields')
+// passing this as a second argument to `router.<verb>` will make it
+// so that a token MUST be passed for that route to be available
+// it will also set `req.user`
+const requireToken = passport.authenticate('bearer', { session: false })
+
+// instantiate a router (mini app that only handles routes)
+const router = express.Router()
+
+///////////////////
+// INDEX 
+// GET (/activities)
+//////////////////
+router.get('/activities', (req, res, next) => {
+    Activity.find()
+    .populate('owner')
+        .then(activities => {
+            return activities.map(activity => activity)
+        })
+        .then(activities => {
+            res.status(200).json({ activities: activities })
+        })
+        .catch(next)
+})
+
+
+///////////////////
+// SHOW 
+// GET (/activities/:id)
+//////////////////
+
+
+
+///////////////////
+// CREATE
+// POST (/activities)
+//////////////////
+
+
+
+///////////////////
+// UPDATE
+// PATCH (/activities/:id)
+//////////////////
+
+
+
+
+///////////////////
+// DESTROY
+// DELETE (/activities/:id)
+//////////////////
+
+module.exports = router
