+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "express-auth-mongoose-boilerplate",
+  "name": "honey-badgers-api",
   "version": "1.0.0",
   "description": "Auth boilerplate for mongoose express using token auth",
   "main": "server.js",
@@ -7,9 +7,10 @@
     "start": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "Timm Schoenborn",
+  "author": "Lauren Routier",
   "license": "ISC",
   "dependencies": {
+    "axios": "^1.1.3",
     "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
